data/youtube/requirements.txt
# Python packages for YouTube scraping
google-api-python-client>=2.100.0
pandas>=2.0.0
python-dotenv>=1.0.0
